refactor(backend): add darwin no-cgo tray fallback
